Remove dead computeTopPicks, drop unreachable null guard, move ClosedPickResultEntry to ranking-score, add rank-skip test
- Remove exported `computeTopPicks` function and its tests from ranking-score; `computeRankedResults` is the sole authoritative scorer
- Remove unreachable `if (!item) continue` guard in the rank-assignment loop (sorted is always Option[])
- Move `ClosedPickResultEntry` interface into ranking-score.ts (primary consumer) and delete computePickResults.ts; update all six import sites
- Add test asserting rank-skip behavior after a tie (scores 4,3,3,2 → ranks 1,2,2,4)
