+++ src/app/groups/[id]/categories/[categoryId]/picks/[pickId]/ClosedPickResultsView.spec.tsx
@@ -1,7 +1,7 @@
 import { cleanup, fireEvent, render, screen } from "@testing-library/react";
 import { afterEach, describe, expect, it, vi } from "vitest";
 
-import type { ClosedPickResultEntry } from "@/lib/computePickResults";
+import type { ClosedPickResultEntry } from "@/lib/ranking-score";
 import type { Option } from "@/lib/types/option";
 
 import { ClosedPickResultsView } from "./ClosedPickResultsView";
