+++ src/app/groups/[id]/categories/[categoryId]/picks/[pickId]/ClosedPickResultsView.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/nextjs-vite";
 
-import type { ClosedPickResultEntry } from "@/lib/computePickResults";
+import type { ClosedPickResultEntry } from "@/lib/ranking-score";
 import type { Option } from "@/lib/types/option";
 
 import { ClosedPickResultsView } from "./ClosedPickResultsView";
