--- src/lib/computePickResults.ts
@@ -1,7 +1,0 @@
-import type { Option } from "@/lib/types/option";
-
-export interface ClosedPickResultEntry {
-  option: Option;
-  rank: number;
-  score: number;
-}
